remove GetSecurityContext class, instead make the getUserName and getUserEmail in base controller.
