+++ src/main/java/gov/usgs/wma/mlrgateway/controller/ExportWorkflowController.java
@@ -19,7 +19,6 @@ import gov.usgs.wma.mlrgateway.StepReport;
 import gov.usgs.wma.mlrgateway.workflow.ExportWorkflowService;
 import gov.usgs.wma.mlrgateway.service.FileExportService;
 import gov.usgs.wma.mlrgateway.service.NotificationService;
-import gov.usgs.wma.mlrgateway.util.GetSecurityContext;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
@@ -31,7 +30,6 @@ public class ExportWorkflowController extends BaseController {
 	private ExportWorkflowService export;
 	public static final String COMPLETE_WORKFLOW = "Complete Export Workflow";
 	public static final String EXPORT_WORKFLOW_SUBJECT = "Transaction File Generation for Requested Location";
-	private GetSecurityContext getSecurityContext = new GetSecurityContext();
 	
 	@Autowired
 	public ExportWorkflowController(ExportWorkflowService export, NotificationService notificationService) {
@@ -49,7 +47,7 @@ public class ExportWorkflowController extends BaseController {
 	public GatewayReport exportWorkflow(@PathVariable("agencyCode") String agencyCode, @PathVariable("siteNumber") String siteNumber, HttpServletResponse response) {
 		setReport(new GatewayReport(COMPLETE_WORKFLOW
 				,null
-				,getSecurityContext.getUserName()
+				,getUserName()
 				,Instant.now().toString()));
 		try {
 			export.exportWorkflow(agencyCode, siteNumber);
