+++ src/main/java/gov/usgs/wma/mlrgateway/controller/WorkflowController.java
@@ -21,7 +21,6 @@ import gov.usgs.wma.mlrgateway.StepReport;
 import gov.usgs.wma.mlrgateway.UserSummaryReport;
 import gov.usgs.wma.mlrgateway.workflow.LegacyWorkflowService;
 import gov.usgs.wma.mlrgateway.service.NotificationService;
-import gov.usgs.wma.mlrgateway.util.GetSecurityContext;
 import gov.usgs.wma.mlrgateway.util.UserSummaryReportBuilder;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -36,7 +35,6 @@ public class WorkflowController extends BaseController {
 	private UserSummaryReportBuilder userSummaryReportbuilder;
 	public static final String COMPLETE_WORKFLOW_SUBJECT = "Submitted Ddot Transaction";
 	public static final String VALIDATE_DDOT_WORKFLOW_SUBJECT = "Submitted Ddot Validation";
-	private GetSecurityContext getSecurityContext = new GetSecurityContext();
 	
 	@Autowired
 	public WorkflowController(LegacyWorkflowService legacy, NotificationService notificationService) {
@@ -54,7 +52,7 @@ public class WorkflowController extends BaseController {
 	public UserSummaryReport legacyWorkflow(@RequestPart MultipartFile file, HttpServletResponse response) {
 		setReport(new GatewayReport(LegacyWorkflowService.COMPLETE_WORKFLOW
 				,file.getOriginalFilename()
-				,getSecurityContext.getUserName()
+				,getUserName()
 				,Instant.now().toString()));
 		userSummaryReportbuilder = new UserSummaryReportBuilder();
 		try {
@@ -91,7 +89,7 @@ public class WorkflowController extends BaseController {
 	public UserSummaryReport legacyValidationWorkflow(@RequestPart MultipartFile file, HttpServletResponse response) {
 		setReport(new GatewayReport(LegacyWorkflowService.VALIDATE_DDOT_WORKFLOW
 				,file.getOriginalFilename()
-				,getSecurityContext.getUserName()
+				,getUserName()
 				,Instant.now().toString()));
 		userSummaryReportbuilder = new UserSummaryReportBuilder();
 		try {
